Use one single search for on display search.
